+++ test/3_eventemitter_websockets_embedded_sanity.js
@@ -39,10 +39,11 @@ describe('3_eventemitter_websockets_embedded_sanity', function () {
   });
 
   after(function (done) {
-    publisherclient.disconnect()
-      .then(listenerclient.disconnect()
-        .then(happnInstance.stop()
-          .then(done)))
+
+    publisherclient.disconnect({timeout:2000})
+      .then(listenerclient.disconnect({timeout:2000}))
+      .then(happnInstance.stop())
+      .then(done)
       .catch(done);
   });