+++ test/7a_websockets_listeners.js
@@ -25,11 +25,11 @@ describe('7a_websockets_listeners', function () {
   after(function (done) {
     this.timeout(10000);
 
-    publisherclient.disconnect(function(e){
-      listenerclient.disconnect(function(e){
-        happnInstance.stop(done);
-      })
-    })
+    publisherclient.disconnect({timeout:2000})
+      .then(listenerclient.disconnect({timeout:2000}))
+      .then(happnInstance.stop())
+      .then(done)
+      .catch(done);
   });
 
   before('should initialize the service', function (callback) {